{{cookiecutter.repo_name}}/requirements.txt
click>=6.7
pyyaml>=3.12
